+++ lib/language_helper.dart
@@ -4,27 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:language_code/language_code.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
+export 'package:language_code/language_code.dart' show LanguageCodes;
+
+part 'src/data_types.dart';
+part 'src/extensions.dart';
 part 'src/language_builder.dart';
-part 'src/language_codes.dart';
 part 'src/language_helper.dart';
-
-typedef LanguageData = Map<LanguageCodes, Map<String, String>>;
-
-extension LanguageHelperEx on String {
-  /// Translate
-  String get tr => LanguageHelper.instance.translate(this);
-
-  /// Translate with parammeters
-  /// ``` dart
-  /// final text = 'result is @{param}'.trP({'param' : 'zero'});
-  /// print(text); // -> 'result is zero'
-  String trP([Map<String, dynamic> params = const {}]) {
-    return LanguageHelper.instance.translate(this, params: params);
-  }
-
-  /// Full version of the translation, includes all parameters.
-  String trF({Map<String, dynamic> params = const {}, LanguageCodes? toCode}) {
-    return LanguageHelper.instance
-        .translate(this, params: params, toCode: toCode);
-  }
-}