+++ lib/src/data_types.dart
@@ -1,0 +1,4 @@
+part of '../language_helper.dart';
+
+/// Data for the [LanguageHelper] to initialize.
+typedef LanguageData = Map<LanguageCodes, Map<String, String>>;